% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle-utils.R
\name{download_chronicle}
\alias{download_chronicle}
\title{Download Duke Chronicle's ZIP file of JPGs}
\usage{
download_chronicle(id, outputdir=".", overwrite=FALSE, filename=NULL)
}
\arguments{
\item{id}{identifier of the chronicle issue. Valid identifiers are under the column "local_id" in the \code{chronicle_meta} data file.}

\item{outputdir}{output directory for the ZIP file.}

\item{overwrite}{whether or not to overwrite data which has already been downloaded.}

\item{filename}{name of the ZIP file. Default is NULL, in which case the ZIP file will be named after the issue id.}
}
\description{
Download ZIP file of JPG newspaper scans for the specified edition. All Chronicle issues are listed in the \code{chronicle_meta} data file with identifiers under the "local_id" column.
}
