% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr-utils.R
\name{hocr_from_images}
\alias{hocr_from_images}
\title{Runs tesseract-OCR on the given image files.}
\usage{
hocr_from_images(imgfiles, outputdir=here(), silent=FALSE, options="")
}
\arguments{
\item{imgfiles}{list of image file paths. Image format should be compatible with tesseract as configured on the user's system. JPEG and PNG files are generally compatible.}

\item{outputdir}{directory where to store the output hocr files.}

\item{silent}{whether or not to supress messages (default is FALSE).}

\item{options}{additional options to pass to the tesseract command line tool. E.g. options="--psm 1" will use page segmentation mode one.}
}
\value{
List of output hocr file paths.
}
\description{
This function performs a system call to the tesseract command line tool, running tesseract-OCR on the given images.
}
