% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr-utils.R
\name{hocr_from_zip}
\alias{hocr_from_zip}
\title{Run tesseract OCR on a zipped set of images.}
\usage{
hocr_from_zip(
  zipped,
  outputdir = ".",
  exdir = NULL,
  silent = FALSE,
  options = ""
)
}
\arguments{
\item{zipped}{path to zip file containing images.}

\item{outputdir}{directory where to store the output hocr files.}

\item{exdir}{directory where images are extracted to. Set to NULL to extract image in a temporary folder.}

\item{silent}{whether or not to supress messages (default is FALSE).}

\item{options}{additional options to pass to the tesseract command line tool. E.g. options="--psm 1" will use page segmentation mode one.}
}
\value{
List of output hocr file paths.
}
\description{
This function unzips images and performs a system call to the tesseract command line tool, running tesseract-OCR on the given images.
}
