% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr-utils.R
\name{visualize_html}
\alias{visualize_html}
\title{In-browser visualization of hocr files}
\usage{
visualize_html(hocr_files, outputdir=".")
}
\arguments{
\item{hocr_files}{list of  hocr file paths to visualize.}

\item{outputdir}{directory where to store the viewable html. This html file has the same name as the hocr file and can be opened in any browser.}
}
\description{
This function creates an html file which can be opened in the browser and which showcases the hocr output using hocrjs.
}
